# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Devops API Sec",
        "title": "NEW Devops API",
        "contact": {},
        "version": "v0.1.0"
    },
    "host": "localhost:4949",
    "paths": {
        "/v1/events/{id}": {
            "post": {
                "description": "HandleEvent endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "HandleEvent endpoint",
                "operationId": "HandleEvent",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/info": {
            "get": {
                "description": "HandleInfo endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "HandleInfo endpoint",
                "operationId": "HandleInfo",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Info"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Action": {
            "type": "object",
            "properties": {
                "key_binding": {
                    "type": "string"
                },
                "name": {
                    "description": "Name is the name to be shown on UI",
                    "type": "string"
                },
                "output_type": {
                    "description": "Output type can be",
                    "type": "string"
                },
                "schema": {
                    "type": "object",
                    "additionalProperties": true
                },
                "type": {
                    "description": "Type can be one of normal, special, internal",
                    "type": "string"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "model.EventResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "result": {}
            }
        },
        "model.Info": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Action"
                    }
                },
                "defaultIsolator": {
                    "type": "string"
                },
                "general": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "isolatorType": {
                    "type": "string"
                },
                "plugins": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "resourceTypes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}